Add support for full uris in the path. This may change in the future. For now the full uris are encoded as '<' + URLEncode(uri) + '>'
Also make sure that all matching is done at the uri level. I'm starting to think that I need to expand the Matcher api to allow the provision of custom word matching. For now to support this I've added a CopyVisitor that provides extension points to allow you to override specific AST node copy semantics.

TS-296 Update the query string builders to allow the selection of partial responses