Make it easier to use partial response with jackson+jaxrs.
This adds a new utility to the jackson module: JacksonFilters - which can be used to configure an ObjectMapper to filter based on a partial response pattern.

I've also added a new module that adds jax-rs specific code to interact with the jackson-jaxrs module, providing a ContextResolver for ObjectMapper that looks for a request parameter and automatically applies the partial response filter for you.

Fixes #5